feat: panic-unwind
This change extracts the kernel's panic.rs logic into a standalone crate (`panic-unwind`) so it can be shared with other library crates such as `fiber`.
