+++ .gitignore
@@ -2,3 +2,4 @@
 .idea
 .DS_Store
 libs/cranelift-codegen/meta/target
+.jj
