+++ kernel/src/shell.rs
@@ -220,7 +220,7 @@ fn handle_command<'cmd>(ctx: Context<'cmd>, commands: &'cmd [Command]) -> CmdRes
         if let Some(current) = chunk.strip_prefix(cmd.name) {
             let current = current.trim();
 
-            return crate::panic::catch_unwind(|| cmd.run(Context { current, ..ctx })).unwrap_or({
+            return panic_unwind::catch_unwind(|| cmd.run(Context { current, ..ctx })).unwrap_or({
                 Err(Error {
                     line: cmd.name,
                     kind: ErrorKind::Other("command failed"),
