+++ kernel/src/wasm/trap_handler.rs
@@ -138,7 +138,7 @@ where
 
         match unwind_reason {
             UnwindReason::Trap(reason) => Err(Trap { reason, backtrace }),
-            UnwindReason::Panic(payload) => crate::panic::resume_unwind(payload),
+            UnwindReason::Panic(payload) => panic_unwind::resume_unwind(payload),
         }
     }
 }
@@ -382,7 +382,7 @@ where
             Err(reason) => (T::SENTINEL, Some(UnwindReason::Trap(reason.into()))),
         };
 
-        crate::panic::catch_unwind(AssertUnwindSafe(f))
+        panic_unwind::catch_unwind(AssertUnwindSafe(f))
             .unwrap_or_else(|payload| (T::SENTINEL, Some(UnwindReason::Panic(payload))))
     }
 }
