+++ libs/panic-unwind/Cargo.toml
@@ -1,0 +1,22 @@
+[package]
+name = "panic-unwind"
+version.workspace = true
+edition.workspace = true
+authors.workspace = true
+license.workspace = true
+
+[dependencies]
+spin.workspace = true
+cpu-local.workspace = true
+unwind2.workspace = true
+abort.workspace = true
+
+# 3rd-party dependencies
+tracing.workspace = true
+#cfg-if.workspace = true
+
+[target.'cfg(any(target_arch = "riscv64", target_arch = "riscv32"))'.dependencies]
+riscv.workspace = true
+
+[lints]
+workspace = true
